+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Feat
 - **alpha:** fortune only called if it exist
+- **catppuccin:** load from setup now
 - **coq:** <c-h> now does not escape to normal mode first
 - **lsp_lines:** changed packer declaration using as
 - **lsp_lines:** added lsp_lines